Create README.md
This is a Python project, which you can play a stickman to fight a Nazi on Mars. By the way, this is just a little thing I created in my spare time. It's mainly for making you laugh. Please don't bring in any emotions. Certainly, I will update more function in the game in the future.